+++ Sources/TableSection.swift
@@ -88,6 +88,13 @@ open class TableSection {
         rows[index] = row
     }
     
+    open func delete(rowAt index: Int) {
+        rows.remove(at: index)
+    }
+    
+    // MARK: - deprecated methods -
+    
+    @available(*, deprecated, message: "Use 'delete(rowAt:)' method instead")
     open func delete(index: Int) {
         rows.remove(at: index)
     }