+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
-## [2.0.0](https://github.com/maxsokolov/TableKit/releases/tag/1.4.0)
-Released on 2016-09-06. Breaking changes in 2.0.0:
+## [2.1.0](https://github.com/maxsokolov/TableKit/releases/tag/1.4.0)
+Released on 2016-10-19.
+- `action` method was deprecated on TableRow. Use `on` instead.
+- Support multiple actions with same type on row.
+- You could now build your own cell height calculating strategy. See [TablePrototypeCellHeightCalculator](Sources/TablePrototypeCellHeightCalculator.swift).
+- Default distance between sections changed to `UITableViewAutomaticDimension`. You can customize it, see [TableSection](Sources/TableSection.swift)
+
+## [2.0.0](https://github.com/maxsokolov/TableKit/releases/tag/2.0.0)
+Released on 2016-10-06. Breaking changes in 2.0.0:
 <br/>The signatures of `TableRow` and `TableRowAction` classes were changed from
 ```swift
 let action = TableRowAction<String, StringTableViewCell>(.click) { (data) in